Conflicts:
	sql/dd/dd_table.cc
	sql/dd/dd_table.h
	sql/dd/dd_tablespace.cc
	sql/dd/dd_tablespace.h
	sql/dd/impl/types/tablespace_impl.cc
	sql/dd_table_share.h
	sql/handler.h
	sql/sql_db.cc
	sql/sql_partition_admin.cc
	sql/sql_rename.cc
	sql/sql_table.cc
	sql/sql_table.h
	sql/sql_tablespace.cc
	sql/sql_truncate.cc
	sql/sql_truncate.h
	sql/table.cc

+++ sql/dd/dd_upgrade.cc
@@ -13,9 +13,23 @@
    along with this program; if not, write to the Free Software Foundation,
    51 Franklin Street, Suite 500, Boston, MA 02110-1335 USA */
 
-#include "dd_upgrade.h"
-#include "dd/impl/bootstrapper.h"             // execute_query
+#include <memory>
+#include <vector>
+#include <string>
 
+#include "dd/cache/dictionary_client.h"       // dd::cache::Dictionary_client
+#include "dd/dd_event.h"                      // create_event
+#include "dd/dd.h"                            // dd::get_dictionary
+#include "dd/dd_schema.h"                     // create_schema
+#include "dd/dd_table.h"                      // create_dd_user_table
+#include "dd/dd_tablespace.h"                 // create_tablespace
+#include "dd/dd_trigger.h"                    // dd::create_trigger
+#include "dd/dd_view.h"                       // create_view
+#include "dd/impl/bootstrapper.h"             // execute_query
+#include "dd/impl/dictionary_impl.h"          // dd::Dictionary_impl
+#include "dd/types/object_type.h"             // dd::Object_type
+#include "dd/types/tablespace.h"              // dd::Tablespace
+#include "dd_upgrade.h"
 #include "derror.h"                           // ER_DEFAULT
 #include "event_db_repository.h"              // Events
 #include "event_parse_data.h"                 // Event_parse_data
@@ -23,51 +37,35 @@
 #include "lock.h"                             // Tablespace_hash_set
 #include "log.h"                              // sql_print_warning
 #include "mysqld.h"                           // mysql_real_data_home
+#include "mysql/psi/mysql_file.h"             // mysql_file_open
 #include "my_user.h"                          // parse_user
 #include "parse_file.h"                       // File_option
 #include "partition_info.h"                   // partition_info
-#include "sql_partition.h"                    // mysql_unpack_partition
-#include "sql_table.h"                        // build_tablename
-#include <mysql/psi/mysql_file.h>             // mysql_file_open
-#include <psi_memory_key.h>                   // key_memory_TABLE
-#include "sql_base.h"                         // open_tables
-#include "sql_parse.h"                        // check_string_char_length
-#include "sql_plugin.h"                       // plugin_unlock
+#include "psi_memory_key.h"                   // key_memory_TABLE
 #include "sp.h"                               // db_load_routine
 #include "sp_head.h"                          // sp_head
+#include "sql_base.h"                         // open_tables
 #include "sql_lex.h"                          // new_empty_query_block
+#include "sql_parse.h"                        // check_string_char_length
+#include "sql_partition.h"                    // mysql_unpack_partition
+#include "sql_plugin.h"                       // plugin_unlock
 #include "sql_prepare.h"                      // Ed_connection
 #include "sql_show.h"                         // view_store_options
+#include "sql_table.h"                        // build_tablename
 #include "sql_time.h"                         // interval_type_to_name
 #include "sql_view.h"                         // mysql_create_view
+#include "table.h"                            // Table_check_intact
+#include "table_trigger_dispatcher.h"         // Table_trigger_dispatcher
 #include "transaction.h"                      // trans_commit
 #include "trigger.h"                          // Trigger
-#include "table_trigger_dispatcher.h"         // Table_trigger_dispatcher
 #include "tztime.h"                           // my_tz_find
-#include "table.h"                            // Table_check_intact
-
-#include "dd/dd.h"                            // dd::get_dictionary
-#include "dd/dd_event.h"                      // create_event
-#include "dd/dd_schema.h"                     // create_schema
-#include "dd/dd_table.h"                      // create_dd_user_table
-#include "dd/dd_tablespace.h"                 // create_tablespace
-#include "dd/dd_trigger.h"                    // dd::create_trigger
-#include "dd/dd_view.h"                       // create_view
-#include "dd/cache/dictionary_client.h"       // dd::cache::Dictionary_client
-#include "dd/impl/dictionary_impl.h"          // dd::Dictionary_impl
-#include "dd/types/object_type.h"             // dd::Object_type
-#include "dd/types/tablespace.h"              // dd::Tablespace
-
-#include <memory>
-#include <vector>
-#include <string>
 
 namespace dd {
 
-const std::string ISL_EXT= ".isl";
-const std::string PAR_EXT= ".par";
-const std::string OPT_EXT= ".opt";
-const std::string SDI_EXT= ".SDI";
+const String_type ISL_EXT= ".isl";
+const String_type PAR_EXT= ".par";
+const String_type OPT_EXT= ".opt";
+const String_type SDI_EXT= ".SDI";
 const char *TRN_EXT= ".TRN";
 const char *TRG_EXT= ".TRG";
 
@@ -349,17 +347,6 @@ TABLE_FIELD_TYPE proc_table_fields_old[MYSQL_PROC_FIELD_COUNT] =
 
 static const TABLE_FIELD_DEF
   proc_table_def_old= {MYSQL_PROC_FIELD_COUNT, proc_table_fields_old};
-
-
-const std::vector<std::string> dd_table_names=
-    { "version", "character_sets", "collations", "tablespaces",
-      "tablespace_files", "catalogs", "schemata",
-      "st_spatial_reference_systems", "tables", "view_table_usage",
-      "view_routine_usage", "columns", "indexes", "index_column_usage",
-      "column_type_elements", "foreign_keys", "foreign_key_column_usage",
-      "table_partitions", "table_partition_values", "index_partitions",
-      "events", "routines", "parameters", "parameter_type_elements",
-      "triggers", "table_stats", "index_stats" };
 
 
 /**
@@ -839,7 +826,7 @@ class Upgrade_MDL_guard
 
   THD *m_thd;
 public:
-  bool acquire_lock(const std::string &db_name, const std::string &table_name)
+  bool acquire_lock(const String_type &db_name, const String_type &table_name)
   {
     return dd::acquire_exclusive_schema_mdl(m_thd, db_name.c_str(),
                                             false, &m_mdl_ticket_schema) ||
@@ -1002,7 +989,7 @@ public:
       Make a copy of mem_root as TABLE object is allocated within its
       own mem_root and free_root() updates its argument.
     */
-    MEM_ROOT m_root= *m_mem_root;
+    MEM_ROOT m_root= std::move(*m_mem_root);
     free_root(&m_root, MYF(0));
   }
 };
@@ -1112,8 +1099,8 @@ static File_option view_parameters[]=
 */
 static bool create_unlinked_view(THD *thd,
                                  TABLE_LIST *view_ref,
-                                 const std::string &db_name,
-                                 const std::string &view_name)
+                                 const String_type &db_name,
+                                 const String_type &view_name)
 {
   SELECT_LEX *backup_select= thd->lex->select_lex;
   TABLE_LIST *saved_query_tables= thd->lex->query_tables;
@@ -1151,12 +1138,14 @@ static bool create_unlinked_view(THD *thd,
   @param[out] str       String object to store view definition.
   @param[in]  db_name   database name.
   @param[in]  view_name view name.
+  @param[in]  cs        Charset Information.
 */
 static void create_alter_view_stmt(THD *thd,
                                    TABLE_LIST *view_ref,
                                    String *str,
-                                   const std::string &db_name,
-                                   const std::string &view_name)
+                                   const String_type &db_name,
+                                   const String_type &view_name,
+                                   const CHARSET_INFO *cs)
 {
   str->append(STRING_WITH_LEN("ALTER "));
   view_store_options(thd, view_ref, str);
@@ -1165,7 +1154,7 @@ static void create_alter_view_stmt(THD *thd,
   str->append('.');
   append_identifier(thd, str, view_name.c_str(), view_name.length());
   str->append(STRING_WITH_LEN(" AS "));
-  str->append(view_ref->select_stmt.str, view_ref->select_stmt.length);
+  str->append(view_ref->select_stmt.str, view_ref->select_stmt.length, cs);
   if (view_ref->with_check != VIEW_CHECK_NONE)
   {
     if (view_ref->with_check == VIEW_CHECK_LOCAL)
@@ -1192,8 +1181,8 @@ static void create_alter_view_stmt(THD *thd,
 
 */
 static bool fix_view_cols_and_deps(THD *thd, TABLE_LIST *view_ref,
-                                   const std::string &db_name,
-                                   const std::string &view_name,
+                                   const String_type &db_name,
+                                   const String_type &view_name,
                                    MEM_ROOT *mem_root)
 {
   bool error= false;
@@ -1216,6 +1205,7 @@ static bool fix_view_cols_and_deps(THD *thd, TABLE_LIST *view_ref,
 
   thd->variables.character_set_client= m_client_cs;
   thd->variables.collation_connection= m_connection_cl;
+  thd->update_charset();
 
   MEM_ROOT *m_mem_root= thd->mem_root;
   thd->mem_root= mem_root;
@@ -1225,19 +1215,19 @@ static bool fix_view_cols_and_deps(THD *thd, TABLE_LIST *view_ref,
   thd->variables.sql_mode&= ~(MODE_PIPES_AS_CONCAT | MODE_ANSI_QUOTES |
                               MODE_IGNORE_SPACE | MODE_NO_BACKSLASH_ESCAPES);
 
-  String full_view_definition;
+  String full_view_definition((char *)0, 0, m_connection_cl);
   create_alter_view_stmt(thd, view_ref, &full_view_definition,
-                         db_name, view_name);
+                         db_name, view_name, m_connection_cl);
 
   String db_query;
   db_query.append(STRING_WITH_LEN("USE "));
   append_identifier(thd, &db_query, db_name.c_str(), db_name.length());
 
-  std::string change_db_query(db_query.ptr(), db_query.length());
+  String_type change_db_query(db_query.ptr(), db_query.length());
   error= execute_query(thd, change_db_query);
 
   // Execute ALTER view statement to create the view dependency entry in DD.
-  std::string query(full_view_definition.ptr(), full_view_definition.length());
+  String_type query(full_view_definition.ptr(), full_view_definition.length());
   if (!error)
     error= execute_query(thd, query);
 
@@ -1253,8 +1243,11 @@ static bool fix_view_cols_and_deps(THD *thd, TABLE_LIST *view_ref,
     update_view_status(thd, db_name.c_str(), view_name.c_str(), false);
     error= false;
   }
+
+  // Restore variables
   thd->variables.character_set_client= client_cs;
   thd->variables.collation_connection= cs;
+  thd->update_charset();
   thd->mem_root= m_mem_root;
   thd->variables.sql_mode= saved_mode;
 
@@ -1279,8 +1272,8 @@ static bool fix_view_cols_and_deps(THD *thd, TABLE_LIST *view_ref,
 
 static bool migrate_view_to_dd(THD *thd,
                                const FRM_context &frm_context,
-                               const std::string &db_name,
-                               const std::string &view_name,
+                               const String_type &db_name,
+                               const String_type &view_name,
                                MEM_ROOT *mem_root,
                                bool is_fix_view_cols_and_deps)
 {
@@ -1610,8 +1603,8 @@ static bool fill_partition_info_for_upgrade(THD *thd,
 
 static bool add_triggers_to_table(THD *thd,
                                   TABLE *table,
-                                  const std::string &schema_name,
-                                  const std::string &table_name)
+                                  const String_type &schema_name,
+                                  const String_type &table_name)
 {
   List<::Trigger> m_triggers;
   if (Trigger_loader::trg_file_exists(schema_name.c_str(),
@@ -1796,8 +1789,8 @@ static bool fix_generated_columns_for_upgrade(THD *thd,
 */
 
 static bool migrate_table_to_dd(THD *thd,
-                                const std::string &schema_name,
-                                const std::string &table_name,
+                                const String_type &schema_name,
+                                const String_type &table_name,
                                 bool is_fix_view_cols_and_deps)
 {
   int error= 0;
@@ -3279,7 +3272,7 @@ bool find_files_with_metadata(THD *thd, const char *dbname,
 {
   uint i;
   MY_DIR *a;
-  std::string path;
+  String_type path;
   bool error= false;
 
   path.assign(mysql_real_data_home);
@@ -3292,7 +3285,7 @@ bool find_files_with_metadata(THD *thd, const char *dbname,
   }
   for (i = 0; i < (uint)a->number_off_files; i++)
   {
-    std::string file;
+    String_type file;
 
     file.assign(a->dir_entry[i].name);
     if (file.at(0)  == '.')
@@ -3300,7 +3293,7 @@ bool find_files_with_metadata(THD *thd, const char *dbname,
 
     if(!MY_S_ISDIR(a->dir_entry[i].mystat->st_mode))
     {
-      std::string file_ext;
+      String_type file_ext;
       char schema_name[NAME_LEN + 1];
       char table_name[NAME_LEN + 1];
 
@@ -3365,7 +3358,7 @@ bool find_files_with_metadata(THD *thd, const char *dbname,
   Scans datadir for databases and lists all the database names.
 */
 
-bool find_schema_from_datadir(THD *thd, std::vector<std::string> *db_name)
+bool find_schema_from_datadir(THD *thd, std::vector<String_type> *db_name)
 {
   MY_DIR *a;
   uint i;
@@ -3398,7 +3391,7 @@ bool find_schema_from_datadir(THD *thd, std::vector<std::string> *db_name)
   to backup_metadata_57 folder upgrade upgrade is successful.
 */
 
-static bool check_file_extension(const std::string &extn)
+static bool check_file_extension(const String_type &extn)
 {
   // Check for extensions
   if (extn.size() < 4)
@@ -3421,11 +3414,11 @@ void create_metadata_backup(THD *thd)
 {
   uint i;
   MY_DIR *a, *b;
-  std::string path;
+  String_type path;
   char to_path[FN_REFLEN];
   char from_path[FN_REFLEN];
 
-  std::vector<std::string> db_name;
+  std::vector<String_type> db_name;
 
   (void) execute_query(thd,
                        "RENAME TABLE mysql.proc TO mysql.proc_backup_57");
@@ -3458,7 +3451,7 @@ void create_metadata_backup(THD *thd)
   // Scan all files and folders in data directory.
   for (i = 0; i < (uint) a->number_off_files; i++)
   {
-    std::string file;
+    String_type file;
 
     file.assign(a->dir_entry[i].name);
     if (file.at(0)  == '.')
@@ -3472,7 +3465,7 @@ void create_metadata_backup(THD *thd)
     }
     else
     {
-      std::string file_ext;
+      String_type file_ext;
 
       if (file.size() < 4)
         continue;
@@ -3496,9 +3489,9 @@ void create_metadata_backup(THD *thd)
   }
 
   // Iterate through the databases list
-  for (std::string str: db_name)
+  for (String_type str: db_name)
   {
-    std::string dir_name= str.c_str();
+    String_type dir_name= str.c_str();
     char dir_path[FN_REFLEN];
 
     if (fn_format(dir_path, dir_name.c_str(), path.c_str(),  "",
@@ -3526,13 +3519,13 @@ void create_metadata_backup(THD *thd)
     // Scan all files and folders in data directory.
     for (i = 0; i < (uint) b->number_off_files; i++)
     {
-      std::string file;
+      String_type file;
       file.assign(b->dir_entry[i].name);
 
       if ((file.at(0)  == '.') || (file.size() < 4))
         continue;
 
-      std::string file_ext;
+      String_type file_ext;
       file_ext.assign(file.c_str() + file.size() - 4);
 
        // Get the name without the file extension.
@@ -3565,8 +3558,25 @@ void create_metadata_backup(THD *thd)
 
 bool check_for_dd_tables()
 {
-  for (const std::string &table_name : dd_table_names)
+  // Iterate over DD tables, check .frm files
+  for (System_tables::Const_iterator it= System_tables::instance()->begin();
+       it != System_tables::instance()->end();
+       ++it)
   {
+    String_type table_name= (*it)->entity()->name();
+    String_type schema_name(MYSQL_SCHEMA_NAME.str);
+
+    const System_tables::Types *table_type= System_tables::instance()->
+      find_type(schema_name, table_name);
+
+    bool is_stats_table= (table_type != nullptr) &&
+                         (*table_type != System_tables::Types::CORE);
+    is_stats_table &= (strcmp(table_name.c_str(), "innodb_table_stats") == 0) ||
+                      (strcmp(table_name.c_str(), "innodb_index_stats") == 0);
+
+    if (is_stats_table)
+      continue;
+
     char path[FN_REFLEN+1];
     bool not_used;
     build_table_filename(path, sizeof(path) - 1, "mysql", table_name.c_str(),
@@ -3588,16 +3598,34 @@ bool check_for_dd_tables()
   Drop all Data Dictionary tables and all .SDI files created during upgrade.
 */
 
-void drop_dd_tables_and_sdi_files(THD *thd)
+void drop_dd_tables_and_sdi_files(THD *thd,
+       const System_tables::Const_iterator &last_table)
 {
   uint i, j;
   bool error;
 
   error= execute_query(thd, "SET FOREIGN_KEY_CHECKS= 0");
 
-  for (const std::string &table_name : dd_table_names)
+  // Iterate over DD tables, delete tables
+  for (System_tables::Const_iterator it= System_tables::instance()->begin();
+       it != last_table;
+       ++it)
   {
-    std::string query;
+    String_type table_name= (*it)->entity()->name();
+    String_type schema_name(MYSQL_SCHEMA_NAME.str);
+
+    const System_tables::Types *table_type= System_tables::instance()->
+      find_type(schema_name, table_name);
+
+    bool is_stats_table= (table_type != nullptr) &&
+                         (*table_type != System_tables::Types::CORE);
+    is_stats_table &= (strcmp(table_name.c_str(), "innodb_table_stats") == 0) ||
+                      (strcmp(table_name.c_str(), "innodb_index_stats") == 0);
+
+    if (is_stats_table)
+      continue;
+
+    String_type query;
     query.assign("DROP TABLE mysql.");
     query= query + table_name;
     // Try to delete all DD tables even if error occurs.
@@ -3607,12 +3635,12 @@ void drop_dd_tables_and_sdi_files(THD *thd)
   error|= execute_query(thd, "SET FOREIGN_KEY_CHECKS= 1");
 
   if (error)
-    sql_print_error("Unable to drop the DD unables during clean "
+    sql_print_error("Unable to drop the DD tables during clean "
                      "up after upgrade failure");
 
   // Iterate in data directory and delete all .SDI files
   MY_DIR *a, *b;
-  std::string path;
+  String_type path;
 
   path.assign(mysql_real_data_home);
 
@@ -3626,7 +3654,7 @@ void drop_dd_tables_and_sdi_files(THD *thd)
   // Scan all files and folders in data directory.
   for (i = 0; i < (uint)a->number_off_files; i++)
   {
-    std::string file;
+    String_type file;
 
     file.assign(a->dir_entry[i].name);
     if (file.at(0)  == '.')
@@ -3652,13 +3680,13 @@ void drop_dd_tables_and_sdi_files(THD *thd)
       // Scan all files and folders in data directory.
       for (j = 0; j < (uint)b->number_off_files; j++)
       {
-        std::string file2;
+        String_type file2;
         file2.assign(b->dir_entry[j].name);
 
         if ((file2.at(0)  == '.') || (file2.size() < 4))
           continue;
 
-        std::string file_ext;
+        String_type file_ext;
         file_ext.assign(file2.c_str() + file2.size() - 4);
         if (file_ext.compare(0, 4, SDI_EXT) == 0)
         {
@@ -3678,7 +3706,7 @@ void drop_dd_tables_and_sdi_files(THD *thd)
     else
     {
       // Delete .SDI files in data directory created for schema.
-      std::string file_ext;
+      String_type file_ext;
       if (file.size() < 4)
         continue;
       file_ext.assign(file.c_str() + file.size() - 4);
